+++ app/Http/Controllers/UserProfileController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\UserProfile\UpdateRequest;
 use App\Http\Resources\UserProfileResource;
 use App\Http\Resources\UserProfileResourceCollection;
 use App\Repositories\Contracts\UserProfileRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class UserProfileController extends Controller
 {
@@ -31,9 +32,12 @@ class UserProfileController extends Controller
      *
      * @param IndexRequest $request
      * @return UserProfileResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [UserProfile::class, $request->get('userId')]);
+
         $userProfiles = $this->userProfileRepository->findBy($request->all());
         
         return new UserProfileResourceCollection($userProfiles);
@@ -44,9 +48,12 @@ class UserProfileController extends Controller
      *
      * @param StoreRequest  $request
      * @return UserProfileResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [UserProfile::class, $request->get('userId')]);
+
         $userProfile = $this->userProfileRepository->setUserProfile($request->all());
         
         return new UserProfileResource($userProfile);
@@ -57,9 +64,12 @@ class UserProfileController extends Controller
      *
      * @param UserProfile $userProfile
      * @return UserProfileResource
+     * @throws AuthorizationException
      */
     public function show(UserProfile $userProfile)
     {
+        $this->authorize('show', $userProfile);
+
         return new UserProfileResource($userProfile);
     }
 
@@ -69,9 +79,12 @@ class UserProfileController extends Controller
      * @param UpdateRequest $request
      * @param UserProfile $userProfile
      * @return UserProfileResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, UserProfile $userProfile)
     {
+        $this->authorize('update', $userProfile);
+
         $userProfile = $this->userProfileRepository->update($userProfile,$request->all());
 
         return new UserProfileResource($userProfile);
@@ -82,9 +95,12 @@ class UserProfileController extends Controller
      *
      * @param UserProfile $userProfile
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(UserProfile $userProfile)
     {
+        $this->authorize('destroy', $userProfile);
+
         $this->userProfileRepository->delete($userProfile);
         
         return response()->json(null,203);
