+++ app/Http/Controllers/UserProfilePostController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\UserProfilePost\UpdateRequest;
 use App\Http\Resources\UserProfilePostResource;
 use App\Http\Resources\UserProfilePostResourceCollection;
 use App\Repositories\Contracts\UserProfilePostRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 
 class UserProfilePostController extends Controller
@@ -32,9 +33,12 @@ class UserProfilePostController extends Controller
      *
      * @param IndexRequest $request
      * @return UserProfilePostResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [UserProfilePost::class, $request->get('propertyId')]);
+
         $userProfilePosts = $this->userProfilePostRepository->findBy($request->all());
 
         return new UserProfilePostResourceCollection($userProfilePosts);
@@ -45,9 +49,13 @@ class UserProfilePostController extends Controller
      *
      * @param StoreRequest  $request
      * @return UserProfilePostResource
+     * @throws AuthorizationException
+     *
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [UserProfilePost::class, $request->get('propertyId')]);
+
         $userProfilePost = $this->userProfilePostRepository->save($request->all());
 
         return new UserProfilePostResource($userProfilePost);
@@ -58,9 +66,12 @@ class UserProfilePostController extends Controller
      *
      * @param UserProfilePost $userProfilePost
      * @return UserProfilePostResource
+     * @throws AuthorizationException
      */
     public function show(UserProfilePost $userProfilePost)
     {
+        $this->authorize('show', $userProfilePost);
+
         return new UserProfilePostResource($userProfilePost);
     }
 
@@ -70,9 +81,12 @@ class UserProfilePostController extends Controller
      * @param UpdateRequest $request
      * @param UserProfilePost $userProfilePost
      * @return UserProfilePostResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, UserProfilePost $userProfilePost)
     {
+        $this->authorize('update', $userProfilePost);
+
         $userProfilePost = $this->userProfilePostRepository->update($userProfilePost,$request->all());
 
         return new UserProfilePostResource($userProfilePost);
@@ -83,9 +97,12 @@ class UserProfilePostController extends Controller
      *
      * @param UserProfilePost $userProfilePost
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(UserProfilePost $userProfilePost)
     {
+        $this->authorize('destroy', $userProfilePost);
+
         $this->userProfilePostRepository->delete($userProfilePost);
 
         return response()->json(null,204);
